requirements.txt
python-telegram-bot[job-queue]==21.6
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
